Extracts SerialAdaptors to a separate module
