+++ BoneCentral/Peripherals/CMakeLists.txt
@@ -6,11 +6,13 @@ set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=gnu++11 -Wall -Werror -Wextra -peda
 # Shared Stuff
 
 include_directories(include/Interfaces)
-
 link_directories(../../Drivers/Build)
 
 add_subdirectory(Peripherals.I2CAdaptors)
 include_directories(${I2CAdaptor_Includes})
+
+add_subdirectory(Peripherals.SerialAdaptors)
+include_directories(${SerialAdaptor_Includes})
 
 include_directories(include/Peripherals)
 file(GLOB_RECURSE PERIPHERAL_FILES "src/Peripherals/*.cpp")
@@ -37,9 +39,6 @@ add_test(Test Specs)
 
 # Production
 
-include_directories(include/Drivers)
-file(GLOB_RECURSE DRIVER_FILES "src/Drivers/*.cpp")
-add_library(Wrappers SHARED ${DRIVER_FILES})
 
 include_directories(include/Sockets)
 file(GLOB_RECURSE SOCKET_FILES "src/Sockets/*.cpp")
@@ -49,5 +48,4 @@ include_directories(include/Bootstrap)
 file(GLOB_RECURSE BOOTSTRAP_FILES "src/Bootstrap/*.cpp")
 add_executable(Bootstrap ${BOOTSTRAP_FILES})
 
-target_link_libraries(Bootstrap Peripherals.I2CAdaptors Wrappers Peripherals Sockets)
-target_link_libraries(Wrappers Peripherals)
+target_link_libraries(Bootstrap Peripherals.I2CAdaptors Peripherals.SerialAdaptors Peripherals Sockets)
