+++ BoneCentral/Peripherals/Peripherals.SerialAdaptors/include/Headlights.h
