+++ BoneCentral/Peripherals/Peripherals.SerialAdaptors/include/Serial.h
