+++ BoneCentral/Peripherals/Peripherals.SerialAdaptors/include/SerialFactory.h
@@ -6,10 +6,10 @@
 #define PERIPHERALS_SERIAL_FACTORY_H
 
 #include <IThrusterFactory.h>
-#include <SerialThruster.h>
-#include <EscPower.h>
-#include <Headlights.h>
-#include <Serial.h>
+#include "SerialThruster.h"
+#include "EscPower.h"
+#include "Headlights.h"
+#include "Serial.h"
 
 
 class SerialFactory : public IThrusterFactory {
