+++ BoneCentral/Peripherals/Peripherals.SerialAdaptors/src/Headlights.cpp
@@ -2,7 +2,7 @@
 //
 // Created by Kevin Burgon on 02/01/2016
 
-#include "Headlights.h"
+#include "../include/Headlights.h"
 
 Headlights::Headlights(Serial& serial) : serial_(serial)
 {
